Alternate way of rendering.
Instead of SDL video being fixed at 640x480, it sets the SDL video to the resolution of each image, which allows compatibility with images of different resolutions, and it automatically centers them on screen. But in latest versions of SDL for 3DS linear filtering breaks when changing SDL video. So for now, it doesn't look too good.